Move lib inside Expo project and update imports
