+++ WeedGrowApp/lib/weather/fetchWeather.ts
