+++ WeedGrowApp/lib/weather/getPlantAdvice.ts
